SGT new version added

+++ Assets/DS/Ship Infrastructure/Modules/Engine/EngineStat.cs
@@ -7,6 +7,6 @@ namespace DeepSpace{
     public class EngineStat : ScriptableObject {
         public ModuleType moduleType;
         public int moduleSize;
-        public int maxTrust;
+        public float maxTrust;
     }
 }
